fix: «AI kredit berish» tug'ilganidan beri o'lik edi — SQL parametriga ::int, qorovul butun server/ ga kengaytirildi
Founder `credit_grant` ni admin paneldan birinchi marta jonli bosdi va u
yiqildi: Telegram'da «❌ Bajarilmadi: ichki xato», alertda
`adminAction run xatosi: operator is not unique: unknown + unknown`
(versiya 3636d57).

NUQSON: routes/admin.js -> credit_grant.run dagi `VALUES ($1, $2 + $3, 0)`.
node-postgres parametrlarni TURSIZ yuboradi, ya'ni Postgres uchun ikkalasi
ham `unknown` va u qaysi `+` operatorini tanlashni bilmaydi. Yiqilish PARSE
bosqichida bo'ladi, ya'ni ON CONFLICT DO UPDATE yo'liga umuman yetib
borilmaydi — demak amal tug'ilganidan beri HECH QACHON ishlamagan.

TUZATISH: `VALUES ($1, $2::int + $3::int, 0)` — ai_credits.balance INT
(db/019). DO UPDATE qatoriga tegilmadi: u yerda ustun turi ma'lum va $3
shunga keltiriladi, ya'ni u sog'lom edi.

Tashxis jonli production bazasida qayta hosil qilindi: tipsiz
`PREPARE p1 AS SELECT $1 + $2` -> `ERROR: operator is not unique:
unknown + unknown`, tipli variant -> `PREPARE`; `\d ai_credits` ->
`balance | integer` (db/019 bilan mos, ikki mustaqil manba).

MUHIMROQ QISMI — QOROVUL. Aynan shu naqsh 2026-08-07 da routes/ai.js ->
takeCredits da bo'lgan va unga Test 14o ning 6-bandi yozilgan edi. Lekin u
bitta funksiya tanasini o'qirdi, ya'ni TOPILGAN nuqson kengligida yozilgan
edi, MUMKIN BO'LGAN xato kengligida emas — naqsh boshqa faylda qaytarilganda
test yashil turaverdi va nuqson production'ga chiqdi. Endi qorovul server/,
server/lib/, server/routes/ dagi hamma .js faylni skanerlaydi; ro'yxat
manbadan yig'iladi, test.js ning o'zi chetda, izohlar jsSofi bilan tahlildan
oldin olib tashlanadi.

4 mutatsiya bilan sinaldi, 4 tasi ham to'g'ri javob berdi. Eng muhimi:
naqsh lib/order-history.js ga qo'yilganda ESKI qorovul bilan butun to'plam
YASHIL qolardi — kengaytirishning isboti shu.

91 test yashil (ikki mustaqil usul bilan sanaldi).
Kesh: panel.js 59 -> 60 (Test 16 jadvali birga) — «statik fayl tegilmagan»
da'vosi yarim to'g'ri edi, panelni yangilash ishining o'zi panel.js ga
tegadi va buni Test 16 ushladi.

+++ server/routes/admin.js
@@ -1138,9 +1138,17 @@ const ADMIN_ACTIONS = {
       `<b>Qo'shiladi:</b> +${p.amount}`,
     async run(a) {
       const n = Number(a.payload.amount);
+      // ⚠️ `::int` SHART: turi yozilmagan ikki parametr Postgres uchun
+      // `unknown` bo'ladi va u qaysi `+` operatorini tanlashni bilmaydi
+      // ("operator is not unique: unknown + unknown"). `DO UPDATE` qatorida
+      // kerak emas — u yerda `ai_credits.balance` turi ma'lum va $3 shunga
+      // keltiriladi. Aynan shu nuqson 2026-08-07 da `routes/ai.js` da
+      // bo'lgan, 2026-08-31 da esa SHU YERDA takrorlandi: qorovul (Test 14o)
+      // faqat `takeCredits` ga qarardi, ya'ni topilgan nuqson kengligida
+      // yozilgan edi. Endi qorovul butun `server/` ni skanerlaydi.
       const { rows } = await pool.query(
         `INSERT INTO ai_credits (tg_user_id, balance, spent)
-         VALUES ($1, $2 + $3, 0)
+         VALUES ($1, $2::int + $3::int, 0)
          ON CONFLICT (tg_user_id)
          DO UPDATE SET balance = ai_credits.balance + $3, updated_at = now()
          RETURNING balance`,
